pyproject.toml
[tool.poetry]
name = "cosmos-sdk"
authors = ["Terraform Labs, PTE. <engineering@terra.money>", "Fabio Nukui <fabio.nukui@gmail.com>"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
description = "The Python SDK for Terra"
documentation = "https://terra-money.github.io/terra-sdk-python/index.html"
homepage = "https://github.com/fabio-nukui/cosmos.py"
keywords = ["blockchain", "cosmos", "defi", "crypto", "terra", "osmosis"]
license = "MIT"
packages = [{ include = "cosmos_sdk" }]
readme = "README.md"
repository = "https://github.com/fabio-nukui/cosmos.py"
version = "2.0.0-beta2"

[tool.poetry.dependencies]
aiohttp = "^3.7.3"
attrs = "^21.2.0"
bech32 = "^1.2.0"
bip32utils = "^0.3.post4"
ecdsa = "^0.16.1"
mnemonic = "^0.19"
python = "^3.9"
wrapt = "^1.12.1"
nest-asyncio = "^1.5.1"
protobuf = "^3.19.1"
betterproto = "2.0.0b3"
furl = "^2.1.3"
boltons = "^21.0.0"
cosmos-proto = { url = "https://mg4r5dq4tpgvnm6tu5rwukctke5x3gfw.s3.amazonaws.com/cosmos.proto/python/dist/cosmos_proto-0.1.0-py3-none-any.whl" }

[tool.poetry.dev-dependencies]
black = "^20.8b1"
coverage = "^5.3.1"
flake8 = "^3.8.4"
isort = "^5.7.0"
mypy = "^0.800"
pip = "^21.0"
pytest = "^6.2.1"
pytest-cov = "^2.11.1"
pytest-mock = "^3.5.1"
pytest-runner = "^5.2"
pytest-sugar = "^0.9.4"
watchdog = "^1.0.2"
wheel = "^0.36.2"
Sphinx = "^3.4.3"
sphinx-rtd-theme = "^0.5.1"
sphinx-autodoc-typehints = "^1.11.1"
sphinx-autobuild = "^2020.9.1"
uvloop = "^0.14.0"
aioresponses = "^0.7.2"
asynctest = "^0.13.0"

[tool.pytest.ini_options]
addopts = "-ra -q"
minversion = "6.0"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
